Add processor configuration that synthesizes printf in VCode RoCC
Just like with the VCodeRocketConfig, you just need to specify the
CONFIG parameter in the command line when running `make`.

The synthesized output is put to STDOUT. The `make run-binary` target
will redirect this to the `sims/<sim>/output/CONFIG/test-name.out`
file. If you want to watch the output, use `tail -f <file-name>`.